requirements.txt
Flask==3.0.0
flask_cors==4.0.0
gunicorn==21.2.0
line-bot-sdk==3.5.1
python-dotenv==1.0.0
openai==1.3.8
ipython==8.18.1